+++ packages/core/src/Core.types.ts
@@ -8,8 +8,13 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-import type { ConfigInterface, ConfigurationObject } from "@mocks-server/config";
-import type { LoggerInterface } from "@mocks-server/logger";
+import type {
+  ConfigInterface,
+  ConfigurationObject,
+  GetOptionValueTypeFromDefinition,
+  OptionDefinition,
+} from "@mocks-server/config";
+import type { LoggerInterface, LogLevel } from "@mocks-server/logger";
 import type { Package } from "update-notifier";
 
 import type { AlertsInterface } from "./alerts/types";
@@ -17,6 +22,25 @@ import type { FilesInterface } from "./files/types";
 import type { MockInterface } from "./mock/types";
 import type { ServerInterface } from "./server/types";
 import type { VariantHandlersInterface } from "./variant-handlers/types";
+
+declare global {
+  //eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace MocksServer {
+    //eslint-disable-next-line @typescript-eslint/no-empty-interface
+    interface Config {}
+  }
+}
+
+export type LogOptionDefinition = OptionDefinition<string, { hasDefault: true }>;
+
+declare global {
+  //eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace MocksServer {
+    interface Config {
+      log?: GetOptionValueTypeFromDefinition<LogOptionDefinition, LogLevel>;
+    }
+  }
+}
 
 /** Mocks-server base core interface */
 export interface BaseCoreInterface {