+++ packages/core/src/files/Files.types.ts
@@ -8,7 +8,13 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-import type { ConfigNamespaceInterface } from "@mocks-server/config";
+import type { RegisterOptions } from "@babel/register";
+import type {
+  ConfigNamespaceInterface,
+  OptionDefinition,
+  UnknownObject,
+  GetOptionValueTypeFromDefinition,
+} from "@mocks-server/config";
 import type { LoggerInterface } from "@mocks-server/logger";
 
 import type { AlertsInterface } from "../alerts/types";
@@ -19,6 +25,30 @@ import type {
   FilesLoaderId,
   FilesLoaderOnLoadMethod,
 } from "./FilesLoader.types";
+
+export type FilesEnabledOptionDefinition = OptionDefinition<boolean, { hasDefault: true }>;
+export type FilesPathOptionDefinition = OptionDefinition<string, { hasDefault: true }>;
+export type FilesWatchOptionDefinition = OptionDefinition<boolean, { hasDefault: true }>;
+
+export type BabelEnabledOptionDefinition = OptionDefinition<boolean, { hasDefault: true }>;
+export type BabelConfigOptionDefinition = OptionDefinition<UnknownObject, { hasDefault: true }>;
+
+declare global {
+  //eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace MocksServer {
+    interface Config {
+      files?: {
+        enabled?: GetOptionValueTypeFromDefinition<FilesEnabledOptionDefinition>;
+        path?: GetOptionValueTypeFromDefinition<FilesPathOptionDefinition>;
+        watch?: GetOptionValueTypeFromDefinition<FilesWatchOptionDefinition>;
+        babelRegister?: {
+          enabled?: GetOptionValueTypeFromDefinition<BabelEnabledOptionDefinition>;
+          config?: GetOptionValueTypeFromDefinition<BabelConfigOptionDefinition, RegisterOptions>;
+        };
+      };
+    }
+  }
+}
 
 /** Options for creating a files loader */
 export interface CreateFilesLoaderOptions {