+++ packages/core/src/mock/collections/Collections.types.ts
@@ -8,7 +8,11 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-import type { ConfigNamespaceInterface } from "@mocks-server/config";
+import type {
+  ConfigNamespaceInterface,
+  OptionDefinition,
+  GetOptionValueTypeFromDefinition,
+} from "@mocks-server/config";
 import type { LoggerInterface } from "@mocks-server/logger";
 
 import type { AlertsInterface } from "../../alerts/types";
@@ -21,6 +25,19 @@ import type {
   CollectionPlainObject,
   CollectionPlainObjectLegacy,
 } from "./Collection.types";
+
+export type SelectedCollectionOptionDefinition = OptionDefinition<string>;
+
+declare global {
+  //eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace MocksServer {
+    interface ConfigMockNamespace {
+      collections?: {
+        selected?: GetOptionValueTypeFromDefinition<SelectedCollectionOptionDefinition>;
+      };
+    }
+  }
+}
 
 /** Options for creating a Collections interface */
 export interface CollectionsOptions {