+++ packages/core/src/plugins/Plugins.ts
@@ -8,11 +8,7 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-import type {
-  ConfigNamespaceInterface,
-  OptionInterfaceOfType,
-  OptionDefinition,
-} from "@mocks-server/config";
+import type { ConfigNamespaceInterface, OptionInterfaceOfType } from "@mocks-server/config";
 import type { LoggerInterface } from "@mocks-server/logger";
 import isPromise from "is-promise";
 import { isFunction } from "lodash";
@@ -31,13 +27,15 @@ import type {
   PluginId,
   PluginWithError,
   PluginLifeCycleMethod,
+  PluginsOptionDefinition,
 } from "./Plugins.types";
 
-const OPTIONS: [OptionDefinition<PluginConstructor[], { hasDefault: true }>] = [
+const OPTIONS: [PluginsOptionDefinition] = [
   {
     description: "Plugins to be registered",
     name: "register",
     type: "array",
+    itemsType: "unknown",
     default: [],
     extraData: {
       scaffold: {