+++ packages/core/src/plugins/Plugins.types.ts
@@ -8,12 +8,30 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-import type { ConfigNamespaceInterface } from "@mocks-server/config";
+import type {
+  ConfigNamespaceInterface,
+  OptionDefinition,
+  GetOptionValueTypeFromDefinition,
+} from "@mocks-server/config";
 import type { LoggerInterface } from "@mocks-server/logger";
 
 import type { AlertsInterface } from "../alerts/types";
 import type { ScopedCoreInterface } from "../common/types";
 import type { CoreInterface } from "../Core.types";
+
+export type PluginsOptionDefinition = OptionDefinition<PluginConstructor[], { hasDefault: true }>;
+
+declare global {
+  //eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace MocksServer {
+    interface PluginsConfig {
+      register?: GetOptionValueTypeFromDefinition<PluginsOptionDefinition, PluginConstructor[]>;
+    }
+    interface Config {
+      plugins?: PluginsConfig;
+    }
+  }
+}
 
 /** Options for creating a plugins interface */
 export interface PluginsOptions {