+++ packages/plugin-openapi/src/Plugin.types.ts
@@ -1,4 +1,25 @@
+import type { OptionDefinition, GetOptionValueTypeFromDefinition } from "@mocks-server/config";
 import type { ScopedCoreInterface, PluginId } from "@mocks-server/core";
+
+export type CollectionIdOptionDefinition = OptionDefinition<
+  string,
+  { hasDefault: true; nullable: true }
+>;
+export type CollectionFromOptionDefinition = OptionDefinition<string>;
+
+declare global {
+  //eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace MocksServer {
+    interface PluginsConfig {
+      openapi?: {
+        collection?: {
+          id?: GetOptionValueTypeFromDefinition<CollectionIdOptionDefinition>;
+          from?: GetOptionValueTypeFromDefinition<CollectionFromOptionDefinition>;
+        };
+      };
+    }
+  }
+}
 
 /** Plugin constructor */
 export interface PluginConstructor {